+++ azext_iot/central/models/__init__.py
@@ -8,12 +8,16 @@ from azext_iot.central.models.devicev1 import DeviceV1
 from azext_iot.central.models.devicetwin import DeviceTwin
 from azext_iot.central.models.templatepreview import TemplatePreview
 from azext_iot.central.models.templatev1 import TemplateV1
+from azext_iot.central.models.deviceGroupPreview import DeviceGroupPreview
+from azext_iot.central.models.rolePreview import RolePreview
 
 
 __all__ = [
     "DevicePreview",
+    "DeviceGroupPreview",
     "DeviceV1",
     "DeviceTwin",
     "TemplatePreview",
     "TemplateV1",
+    "RolePreview"
 ]