+++ azext_iot/tests/utility/test_iot_utility_unit.py
@@ -376,14 +376,15 @@ class TestCliInit(object):
         invalid_directories = []
         for directory in directory_structure:
             if directory_structure[directory] is None:
-                invalid_directories.append("Directory: '{}' missing __init__.py".format(directory))
+                invalid_directories.append(
+                    "Directory: '{}' missing __init__.py".format(directory)
+                )
 
         if invalid_directories:
             pytest.fail(", ".join(invalid_directories))
 
     def test_ensure_azure_namespace_path(self):
         import azure
-
         from azext_iot.common.utility import ensure_azure_namespace_path
         from azure.cli.core.extension import get_extension_path
         from azext_iot.constants import EXTENSION_NAME
@@ -392,6 +393,10 @@ class TestCliInit(object):
         original_sys_path = list(sys.path)
         original_azure_namespace_path = list(azure.__path__)
         ext_azure_dir = os.path.join(ext_path, "azure")
+
+        # TODO: Delete after adding azure.* dependency.
+        if not os.path.exists(ext_azure_dir):
+            os.makedirs(ext_azure_dir)
 
         try:
             ensure_azure_namespace_path()
@@ -404,8 +409,8 @@ class TestCliInit(object):
                 list(azure.__path__)
             sys.path[:] = original_sys_path
 
-        assert original_azure_namespace_path == modified_azure_namespace_path[1:]
-        assert modified_azure_namespace_path[0] == ext_azure_dir
+        assert original_azure_namespace_path == modified_azure_namespace_path[:-1]
+        assert modified_azure_namespace_path[-1] == ext_azure_dir
         assert original_azure_namespace_path == list(azure.__path__)
 
         assert original_sys_path == modified_sys_path[1:]